html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-stable-state-document-write-script.html

<!doctype html>
<title>NO stable state when parsing script end tag using document.write</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  var v = document.createElement('video');
  v.load();
  assert_equals(v.networkState, v.NETWORK_NO_SOURCE);
  window.scriptDidRun = false;
  document.write('<script>scriptDidRun = true;<\/script>');
  assert_true(scriptDidRun);
  assert_equals(v.networkState, v.NETWORK_NO_SOURCE);
});
</script>
